Naviagation follows page order correctly
