--- node/risk-app/server/models/rivers-and-sea.js
@@ -1,9 +1,0 @@
-function riversAndSeaViewModel (riskProbability, address, backLink) {
-  this.easting = address.x
-  this.northing = address.y
-  this.backLink = backLink
-  this.riskProbability = riskProbability
-  this.riskStyle = riskProbability.replace(/ /g, '-')
-}
-
-module.exports = riversAndSeaViewModel
