--- node/risk-app/server/models/surface-water.js
@@ -1,10 +1,0 @@
-function surfaceWaterViewModel (riskProbability, address, llfa, backLink) {
-  this.easting = address.x
-  this.northing = address.y
-  this.backLink = backLink
-  this.riskProbability = riskProbability
-  this.riskStyle = riskProbability.replace(/ /g, '-')
-  this.llfa = llfa
-}
-
-module.exports = surfaceWaterViewModel
