+++ node/risk-app/server/views/partials/flagged/nav-gw.html
@@ -1,0 +1,10 @@
+{% if currentpage === 'groundwater' %}
+<li class="gem-c-contents-list__list-item gem-c-contents-list__list-item--dashed gem-c-contents-list__list-item--active" aria-current="true">
+  <strong>Groundwater and reservoirs</strong>
+ </li>
+{% else %}
+<li class="gem-c-contents-list__list-item gem-c-contents-list__list-item--dashed">
+  <a class="gem-c-contents-list__link govuk-link" href="/ground-water" 
+  aria-label="Ground water and reservoirs understand your flood risk page." previewlistener="true">Groundwater and reservoirs</a>
+</li>
+{% endif %}
