+++ node/risk-app/server/views/partials/flagged/nav-rs.html
@@ -1,0 +1,12 @@
+{% if currentpage === 'rivers-and-sea' %}
+<li
+class="gem-c-contents-list__list-item gem-c-contents-list__list-item--dashed gem-c-contents-list__list-item--active"
+aria-current="true">
+<strong>Rivers and the sea</strong>
+</li>
+{% else %}
+<li class="gem-c-contents-list__list-item gem-c-contents-list__list-item--dashed">
+  <a class="gem-c-contents-list__link govuk-link" href="/rivers-and-sea" previewlistener="true" 
+  aria-label="Rivers and the sea understand your flood risk page.">Rivers and the sea</a>
+</li>
+{% endif %}
