+++ node/risk-app/server/views/partials/flagged/nav-sw.html
@@ -1,0 +1,12 @@
+{% if currentpage === 'surface-water' %}
+<li
+class="gem-c-contents-list__list-item gem-c-contents-list__list-item--dashed gem-c-contents-list__list-item--active"
+aria-current="true">
+<strong>Surface
+  water</strong>
+</li>
+{% else %}
+<li class="gem-c-contents-list__list-item gem-c-contents-list__list-item--dashed">
+  <a class="gem-c-contents-list__link govuk-link" href="/surface-water" previewlistener="true">Surface water</a>
+</li>
+{% endif %}
