+++ node/risk-app/server/views/surface-water.html
@@ -1,7 +1,7 @@
 {% extends 'layout.html' %}
 
 {% set pageTitle = "Surface water: understand your flood risk" %}
-
+{% set currentpage = "surface-water" %}
 
 {% block head %}
 {{ super() }}
@@ -28,21 +28,15 @@
   <nav aria-label="Pages with more detail on this service" class="gem-c-contents-list govuk-!-margin-bottom-3"
   role="navigation">
   <ol class="gem-c-contents-list__list">
-    <li
-      class="gem-c-contents-list__list-item gem-c-contents-list__list-item--dashed gem-c-contents-list__list-item--active"
-      aria-current="true">
-      <strong>Surface
-        water</strong>
-    </li>
-    <li class="gem-c-contents-list__list-item gem-c-contents-list__list-item--dashed">
-      <a class="gem-c-contents-list__link govuk-link" href="/rivers-and-sea" previewlistener="true" 
-      aria-label="Rivers and the sea understand your flood risk page.">Rivers and the sea</a>
-    </li>
-    <li class="gem-c-contents-list__list-item gem-c-contents-list__list-item--dashed">
-      <a class="gem-c-contents-list__link govuk-link" href="/ground-water" 
-      aria-label="Ground water and reservoirs understand your flood risk page." previewlistener="true">Groundwater and reservoirs</a>
-    </li>
-  </ol>
+    {% if surfaceWaterIsFirst %}
+    {% include 'partials/flagged/nav-sw.html' %}
+    {% endif %}
+    {% include 'partials/flagged/nav-rs.html' %}
+    {% if not surfaceWaterIsFirst %}
+    {% include 'partials/flagged/nav-sw.html' %}
+    {% endif %}
+    {% include 'partials/flagged/nav-gw.html' %}
+</ol>
 </nav>
 
 <div class="govuk-grid-row">
@@ -50,8 +44,8 @@
     <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
     <h2 class="govuk-heading-m">Yearly chance of flooding</h2>
     <p class="govuk-body">The yearly chance of surface water flooding is:
-      <strong class="govuk-tag govuk-tag--{{ riskStyle }} label-mod-depth">
-        <span>{{ riskProbability }}</span>
+      <strong class="govuk-tag govuk-tag--{{ surfaceWaterStyle }} label-mod-depth">
+        <span>{{ surfaceWaterRisk }}</span>
       </strong>
     </p>
     <p class="govuk-body">This information tells you the highest chance of flooding on the land 15 metres around a property.</p>
@@ -72,12 +66,13 @@
       <p class="govuk-body">It's a good idea to check your flood risk regularly because the information may change.</p>
       <p class="govuk-body">You can also read about <a href="https://www.gov.uk/prepare-for-flooding">how to prepare yourself and your property</a>.</p>
       <p class="govuk-body">Lead local flood authorities are responsible for managing flood risk from surface water. They may hold more detailed information.</p>
-      <p class="govuk-body">Your lead local flood authority is <strong>{{llfa}}</strong>.</p>
+      <p class="govuk-body">Your lead local flood authority is <strong>{{leadLocalFloodAuthority}}</strong>.</p>
       
     <hr
       class="govuk-section-break govuk-section-break--m govuk-section-break--visible  govuk-!-margin-top-6 govuk-!-margin-bottom-3">
 
     <nav class="govuk-pagination govuk-pagination--block" role="navigation" aria-label="results">
+      {% if surfaceWaterIsFirst %}
       <div class="govuk-pagination__next">
         <a class="govuk-link govuk-pagination__link" href="/rivers-and-sea" rel="next"
           previewlistener="true"> <svg class="govuk-pagination__icon govuk-pagination__icon--next"
@@ -89,6 +84,31 @@
           </svg> <span class="govuk-pagination__link-title">Next</span><span class="govuk-visually-hidden">:</span>
           <span class="govuk-pagination__link-label">Rivers and the sea </span></a>
       </div>
+      {% else %}
+      <nav class="govuk-pagination govuk-pagination--block" role="navigation" aria-label="results">
+        <div class="govuk-pagination__prev">
+          <a class="govuk-link govuk-pagination__link" href="/rivers-and-sea" rel="prev" previewlistener="true">
+            <svg class="govuk-pagination__icon govuk-pagination__icon--prev" xmlns="http://www.w3.org/2000/svg"
+              height="13" width="15" aria-hidden="true" focusable="false" viewBox="0 0 15 13">
+              <path
+                d="m6.5938-0.0078125-6.7266 6.7266 6.7441 6.4062 1.377-1.449-4.1856-3.9768h12.896v-2h-12.984l4.2931-4.293-1.414-1.414z">
+              </path>
+            </svg>
+            <span class="govuk-pagination__link-title">Previous</span><span class="govuk-visually-hidden">:</span>
+            <span class="govuk-pagination__link-label">Rivers and Sea</span></a>
+        </div>
+        <div class="govuk-pagination__next">
+          <a class="govuk-link govuk-pagination__link" href="/ground-water" rel="next"
+            previewlistener="true"> <svg class="govuk-pagination__icon govuk-pagination__icon--next"
+              xmlns="http://www.w3.org/2000/svg" height="13" width="15" aria-hidden="true" focusable="false"
+              viewBox="0 0 15 13">
+              <path
+                d="m8.107-0.0078125-1.4136 1.414 4.2926 4.293h-12.986v2h12.896l-4.1855 3.9766 1.377 1.4492 6.7441-6.4062-6.7246-6.7266z">
+              </path>
+            </svg> <span class="govuk-pagination__link-title">Next</span><span class="govuk-visually-hidden">:</span>
+            <span class="govuk-pagination__link-label">Groundwater and reservoirs</span></a>
+        </div>
+  {% endif %}
     </nav>
   </div>
